fcosKola: allow skipping upgrade tests
In `next-devel` right now, we don't have a previous build yet, so
`run-upgrade` doesn't know what parent build to use. There's a good
argument here that we could use the `testing-devel` stream as the
starting stream for now (and actually, that's something that we should
enhance our CI to cover anyway, even when we *do* have `next-devel`
parent builds, since that mimics what nodes will eventually do).

But I just want to get `next-devel` off the ground for now. The first
step either way is allowing us to skip upgrade testing here so that we
can run it manually with more options as needed.