shwrap: set HOME to WORKSPACE
Normally, we're running unprivileged and so Jenkins' setting of `HOME=/`
is not really helpful when running apps like `cargo` and `go build`
which expect to be able to cache things in the homedir.

If for whatever reason someone does want the default `HOME=/` behaviour,
they can just use the less opinionated `sh` step.