pod: drop hardcoded jnlp container
Instead, let the Kubernetes Jenkins plugin inject the default jnlp
container to use so that we're sure it's always using one that's made
for the given k8s version and plugin version.

This is causing issues on the new cluster where the `sh` step would
hang. It seems to be related to running on new k8s but old Jenkins, and
some kind of conflict with how working directories have changed. For
more details, see:

https://issues.jenkins-ci.org/browse/JENKINS-58777
https://issues.jenkins-ci.org/browse/JENKINS-58766