cosaPod: spawn container with dumb-init
Let's use `dumb-init` by default for cosa pods so that we don't end up
with a swarm of zombies taking up resources.

I have a feeling this might also be contributing to the ulimit
exhaustion we're seeing in CI after the move to the new OCP 4.4 cluster.

See also https://github.com/kubernetes/kubernetes/issues/84210.