Add checkoutToDir helper step
This is a thin wrapper around `checkout` which makes it easy to check
out the repo into a subdirectory instead of directly into the workspace.

This can be useful especially in the context of kola external tests,
which use the directory name as part of the test name and you'd prefer a
short predictable name rather than Jenkins' long workspace name.