buildImage: use untilEach() instead of watch() and add timeout
The way we use `watch`, the closure only runs once and then exits. But
we want to keep watching for changes until the build goes to `Complete`.
Instead, just use `untilEach()` which is nicer for this. Then, we can
just keep watching until the build finishes, failing or not. This is
based on the last example in:

https://github.com/openshift/jenkins-client-plugin#watching-and-waiting-of-course

Also add a build timeout of 2 hours to be safe.