fcosKola: make archived artifact filenames unique
Previously if a Jenkins run executed multiple fcosKola calls it would
run multiple times but there would only be one archived kola.tar.xz at
the end of the run because each run overwrote the previous one. Generate
some unique characters to add to the filename of each uploaded file. Also
allow the caller to pass in a marker that will be added to the uploaded
filenames to make them more identifiable.