kola: simplify tests on non-QEMU or if user passed extraArgs
We only need to worry about basic-qemu-scenarios and running reprovision
tests separately when the platform is QEMU (the default). Loosely detect
if the target platform is something other than QEMU by checking to see
if the user provided platformArgs and if we are running against non-QEMU
then just run all the tests up front and not in separate runs.

Similarly if the user passed in `extraArgs` let's assume they want greater
control and could be passing arguments that conflict with our `--tag`
specifications in the invocations below. Given this assumption it would
be better to just run the extraArgs in a single invocation up front.