{kola,kolaTestIso}: workaround issue with accessDenied from dir()
When using `dir()` in the kola and kolaTestIso jobs we are getting a
permission denied error if the directory is isn't under the env.WORKSPACE
directory.

Let's go back to using a `cd ${cosaDir}` for now to workaround this while
we find better solutions. This `cd ${cosaDir}` will have no effect on
multi-arch where we'll be operating in a remote session anyway.