tree: drop cosaParallelCmds
This was a good idea, but in practice the CoreOS pipeline implements a
more sophisticated version of it which include list splitting. We could
try to upstream that, but in the meantime, this is only used by one
upstream CI right now (cosa's), so we'll just inline it there.

The reason for this is that it uses `cosaCmd` and `cosaDir`, both things
I'm trying to get rid of.