kola: run with parallel even if single stage
I think with 380ffa5 it makes the optimization from 362e995 no longer
required. In it's current form the kola-azure and kola-openstack tests
in the pipeline don't have the kola tests broken out into a separate
stage (they skip the upgrade test) so it's hard to find a failure in
the blue ocean view when one happens.