buildImage: put tarball in directory itself
Currently, we're putting the tarball in `/tmp`, but if we're running on
the Jenkins controller, that's a semi-permanent shared location. Since
our generated tarball has a predictable name, we additionally run the
risk of racing with other `buildImage` calls on the controller and then
passing the wrong `dir.tar` to `oc start-build`.

Fix this by putting the tarball in the workspace itself. That way, it's
properly lifecycled to the job run, and there's no chance of racing with
other jobs.