Add `HOME` workaround to other shwrap helpers
We had it in `shwrap`, but not `shwrapCapture` and `shwrapRc`. We're
hitting an issue right now where `podman` wants to create `$HOME/.ssh`
when using the remote stuff but because we're running unprivileged,
we get

```
Error: failed to connect: ssh: handshake failed: mkdir /.ssh: permission denied
```

We're still trying to determine why this is new, but let's fix the
inconsistency for now.