kola: add support for allowRerunSuccess
We'll start now to default to passing in --allow-rerun-success tags=needs-internet
so that we can get notified less about intermittent network flakes.

If someone wants to disable this they can pass in allowRerunSuccess=false.

If someone wants to override the tags to use for rerun-success they can
use allowRerunSuccessArgs. For example, allowRerunSuccessArgs="tags=all".

Closes https://github.com/coreos/fedora-coreos-pipeline/issues/842