kola: refactor and add support for warn: true tests
This commit refactors the arguments and error handling for `kola`
into a runKola closure to make it easier to manage. Most calls in
this file use mostly the same arguments, so this is a nice cleanup.

This commit also adds the `--on-warn-failure-exit-77` option [1]
and adds handling to mark a test as `warn()` if that happens versus
the typical `error()`.

[1] https://github.com/coreos/coreos-assembler/pull/3551