job.yaml: fix apiVersion and other tweaks
The API version for templates is now `template.openshift.io/v1`.
Possibly `v1` was accepted before but not anymore? Fix it.

While we're here, tweak the template name to make it clearer that the
template itself is never created as an object.

Also remove the parameter default value to hard require clients to
always specify it.